Move config to a new location
